Add asynchronous siteswap validation
The first version of `siteswap` will validate all asynchronous
siteswaps, including vanilla siteswaps, multiplex siteswaps, as well as
theoretical siteswaps with negative throw values, and negative
quantities of throws. The cardinality (number of props) and period of
the pattern are also calculated, and the pattern normalised to convert
all siteswaps to a standard form.